requirements.txt
openai
streamlit_chat
imageio
streamlit_extras
geopandas
folium
streamlit_folium
audio_recorder_streamlit
soundfile
gtts
pyodbc
streamlit_toggle
TTS